+++ lib/devise_otp_authenticatable/hooks/sessions.rb
@@ -24,10 +24,6 @@ module DeviseOtpAuthenticatable::Hooks
         warden.logout
         store_location_for(resource, devise_stored_location) # restore the stored location
         respond_with resource, location: otp_credential_path_for(resource, {challenge: challenge})
-      elsif otp_mandatory_on?(resource) # if mandatory, log in user but send him to the must activate otp
-        set_flash_message(:notice, :signed_in_but_otp) if is_navigational_format?
-        sign_in(resource_name, resource)
-        respond_with resource, location: otp_token_path_for(resource)
       else
         sign_in(resource_name, resource)
         respond_with resource, location: after_sign_in_path_for(resource)
@@ -45,14 +41,5 @@ module DeviseOtpAuthenticatable::Hooks
       resource.otp_enabled && !is_otp_trusted_browser_for?(resource)
     end
 
-    #
-    # the resource -should- have otp turned on, but it isn't
-    #
-    def otp_mandatory_on?(resource)
-      return true if resource.class.otp_mandatory && !resource.otp_enabled
-      return false unless resource.respond_to?(:otp_mandatory)
-
-      resource.otp_mandatory && !resource.otp_enabled
-    end
   end
 end